Bump MessageFlag underlying type to be 64-bit (#4042)

+++ src/widgets/helper/ChannelView.hpp
@@ -27,7 +27,7 @@ using ChannelPtr = std::shared_ptr<Channel>;
 struct Message;
 using MessagePtr = std::shared_ptr<const Message>;
 
-enum class MessageFlag : uint32_t;
+enum class MessageFlag : int64_t;
 using MessageFlags = FlagsEnum<MessageFlag>;
 
 class MessageLayout;
